backend/requirements.txt
accelerate==1.2.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
openai==1.57.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
peft==0.14.0
psutil==6.1.0
pydantic==2.10.3
pydantic_core==2.27.1
PyMuPDF==1.25.0
python-dotenv==1.0.1
python-multipart==0.0.19
pytorch-triton-rocm==3.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.1.1+cpu
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
